+++ bonus/src/menu/settings/settings_buttons.c
@@ -6,11 +6,12 @@
 /*   By: asuc <asuc@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/01 21:50:07 by asuc              #+#    #+#             */
-/*   Updated: 2024/07/01 21:51:35 by asuc             ###   ########.fr       */
+/*   Updated: 2024/07/02 06:05:00 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "_settings_menu.h"
+#include <stdint.h>
 
 static int	is_mouse_over_button(t_vec2i pos_button, t_vec2i size_button,
 		t_vec2i pos_mouse)
@@ -22,7 +23,8 @@ static int	is_mouse_over_button(t_vec2i pos_button, t_vec2i size_button,
 	return (0);
 }
 
-static void	render_button(t_data *data, t_button_info *button, t_vec2i pos_mouse)
+static void	render_button(t_data *data, t_button_info *button,
+		t_vec2i pos_mouse)
 {
 	if (is_mouse_over_button(button->pos, button->size, pos_mouse) == 1)
 	{
@@ -38,18 +40,19 @@ static void	render_button(t_data *data, t_button_info *button, t_vec2i pos_mouse
 			button->pos.y);
 	}
 	mlx_string_put(data->graphics.mlx, data->graphics.win, button->pos.x - 300,
-		button->pos.y + 50, 0xFFFFFFFF, button->label);
+		button->pos.y + 50, (int)(uintptr_t)0xFFFFFFFF, button->label);
 	if (data->graphics.settings.waiting_for_key
 		&& data->graphics.settings.button_state_press == button->press_state)
 	{
 		mlx_string_put(data->graphics.mlx, data->graphics.win, button->pos.x
-			+ 20, button->pos.y + 50, 0xFFFFFFFF, "Press a key");
+			+ 20, button->pos.y + 50, (int)(uintptr_t)0xFFFFFFFF,
+			"Press a key");
 	}
 	else
 	{
 		mlx_string_put(data->graphics.mlx, data->graphics.win, button->pos.x
-			+ 20, button->pos.y + 50, 0xFFFFFFFF,
-			(char *)SDL_GetKeyName(SDL_GetKeyFromScancode(button->key)));
+			+ 20, button->pos.y + 50, (int)(uintptr_t)0xFFFFFFFF,
+			(char *)(uintptr_t)SDL_GetKeyName(SDL_GetKeyFromScancode(button->key)));
 	}
 }
